Part 12 of unit tests update for KRaft
